Firmware builds in bazel
