+++ .bazelrc
@@ -1,2 +1,31 @@
 # Enable toolchain resolution with cc
 build --incompatible_enable_cc_toolchain_resolution
+build --platforms=@bazel_embedded//platforms:cortex_m4_fpu
+
+# Never strip debug symbols
+build --strip=never
+
+build --copt=-Os
+build --copt=-ggdb
+build --copt=-fno-exceptions
+build --copt=-fomit-frame-pointer
+build --copt=-falign-functions=16
+build --cxxopt=-std=c++14
+build --linkopt=-lc
+build --linkopt=-lm
+build --linkopt=-lnosys
+
+# LTO Builds (Do not work with chibios 2.6)
+#build --copt=-ffat-lto-objects
+#build --linkopt=-ffat-lto-objects
+
+# FPU
+build --copt=-mthumb
+build --copt=-fsingle-precision-constant
+build --linkopt=-mthumb
+
+# Code reduction options
+build --copt=-ffunction-sections
+build --copt=-fdata-sections
+build --linkopt=-Wl,--gc-sections
+#build --linkopt=-Wl,--print-gc-sections
