+++ bd_tools/comms.py
@@ -37,8 +37,8 @@ COMM_FLAG_SEND = 0x00
 COMM_FLAG_CRASH = 0x02
 
 COMM_BOOTLOADER_OFFSET = 0x08000000
-COMM_BOARD_ID_OFFSET = 0x08004000
 COMM_NVPARAMS_OFFSET = 0x08008000
+COMM_BOARD_ID_OFFSET = 0x0800C000
 COMM_FIRMWARE_OFFSET = 0x08010000
 COMM_DEFAULT_BAUD_RATE = 1000000
