+++ firmware/proto/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@rules_proto//proto:defs.bzl", "proto_library")
+
+proto_library(
+    name = "messages_proto",
+    srcs = ["messages.proto"],
+    visibility = ["//visibility:public"],
+    deps = ["@com_github_nanopb_nanopb//:nanopb_proto"],
+)
