+++ firmware/src/state.cpp
@@ -1,10 +1,9 @@
 #include "state.hpp"
 
-#include <pb_decode.h>
-#include <pb_encode.h>
-
 #include <cstring>
 
+#include "pb_decode.h"
+#include "pb_encode.h"
 #include "flash.h"
 #include "helper.h"
