+++ toolchains/programs/BUILD
@@ -1,0 +1,15 @@
+package(default_visibility = ["//visibility:public"])
+
+config_setting(
+    name = "firmware",
+    values = {
+        "compilation_mode": "dbg",
+    },
+)
+
+config_setting(
+    name = "bootloader",
+    values = {
+        "compilation_mode": "fastbuild",
+    },
+)
